requirements.txt
huggingface_hub[hf_xet]
hf-xet
timm
mistralai
requests
accelerate
diffusers
invisible_watermark
torch
#xformers
sentencepiece
transformers
peft
psutil
matplotlib
opencv-python
opencv-python-headless 
pillow
openai
google-genai
google-generativeai